Add attestation of accuracy to the issue template
Added an attestation of accuracy to the issue template so that anyone who wishes
to make a contribution knows to conform to the guidelines for contributing as
well as attest to the accuracy of the dimensions being provided if any
conversions were done.

The placeholder description was also removed from the issue template and
replaced with an empty string.